api/requirements.txt
torch
numpy
pillow
requests
torchvision
onnxruntime
azure-functions